Nathan/requirements.txt
# Websockets and async communication
websockets==10.0

# AWS services
boto3==1.20.23
botocore==1.23.23

# Slack integration
slack-sdk==3.11.2

# Database (DynamoDB)
pynamodb==5.2.0

# Serverless framework (if you're using it for AWS Lambda deployment)
serverless==3.2.0

# Testing
pytest==7.2.1
pytest-mock==3.7.3

# For API and HTTP requests
requests==2.26.0

# Utils and other necessary libraries
python-dotenv==0.19.2

# If you are using other frameworks, services or tools, include their libraries here.
